backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
kafka-python==2.0.2
python-dotenv==1.0.1
